% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-action-split.R
\name{add_split}
\alias{add_split}
\alias{remove_split}
\alias{update_split}
\title{Add a split specification to a workflow}
\usage{
add_split(x, .f, ...)

remove_split(x)

update_split(x, .f, ...)
}
\arguments{
\item{x}{A workflow}

\item{.f}{A function to be applied to the dataset in the workflow. Must
return an object of class \code{rsplit}. See package
\code{\link[rsample]{rsample}}.}

\item{...}{arguments passed to \code{.f}. These arguments must be named.
The processing of \code{...} respects the quotation rules from \code{.f}.
In other words, if the function allows variables as strings \strong{and}
as names, the user can specify both. See the example sections.}
}
\value{
\code{x}, updated with either a new or removed split specification.
}
\description{
\itemize{
\item \code{add_split()} specifies the type of splitting used in the analysis. It
accepts a function \code{.f} that will be applied to the data. Only
functions which return an \code{rsplit} object will be allowed. See
package \code{\link[rsample]{rsample}} and the details section. If a model
has been fit before adding the split, it will need to be refit.
\item \code{remove_split()} removes the split specification from the workflow. Note
that it keeps other preprocessing steps such as the recipe.
\item \code{update_split()} first removes the split, then adds a new split
specification. Any model that has already been fit based on this
split will need to be refit.
}
}
\details{
The split specification is an optional step in the workflow. You can add a
dataframe, prepare a recipe and fit the model without splitting into
training/testing.

When applied to the data, the function \code{.f} must return an object
of class \code{rsplit}. These are functions which come from the
\code{\link[rsample]{rsample}} package.
}
\examples{
library(rsample)

wf <- workflow()
wf <- add_data(wf, mtcars)

# Strata as string
wf <- add_split(wf, initial_split, prop = 0.8, strata = "cyl")

wf

# Strata as unquoted name
wf <- update_split(wf, initial_split, prop = 0.8, strata = cyl)

wf

remove_split(wf)

# New split function
update_split(wf, initial_time_split)

}
