% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters.tidyflow}
\alias{parameters.tidyflow}
\title{Extract the parameters of a tidyflow}
\usage{
parameters.tidyflow(x, ...)
}
\arguments{
\item{x}{A data frame or tibble used to begin the tidyflow. This is
optional as the data can be specified with \code{\link[=plug_data]{plug_data()}}.}

\item{...}{Not used.}
}
\value{
A tibble of class parameters
}
\description{
Extract the parameters of a tidyflow
}
\examples{
## TODO
5 + 5

}
