% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-raw_data.R
\name{plug_data}
\alias{plug_data}
\alias{drop_data}
\alias{replace_data}
\title{Add a data to a tidyflow}
\usage{
plug_data(x, data, ...)

drop_data(x)

replace_data(x, data, ...)
}
\arguments{
\item{x}{A tidyflow}

\item{data}{A data frame or tibble.}

\item{...}{Not used.}
}
\value{
\code{x}, updated with either a new or removed data frame.
}
\description{
\itemize{
\item \code{plug_data()} specifies the data used in the analysis. This data is used
for the preprocessing and fitting the model.
\item \code{drop_data()} removes the data from the tidyflow.
\item \code{replace_data()} first removes the data, then replaces the previous
data with the new one. Any model that has already been fit based on this
data will need to be refit.
}
}
\details{
\code{plug_data()} is a required step to construct a minimal tidyflow. We
advise that the data passed to the \code{tidyflow} has already been tested
with the recipe before hand. The \code{tidyflow} is not an ideal workflow
for prototyping with the recipe rather for prototyping with the
model/grid/resample/split.
}
\examples{

wf <- tidyflow()
wf <- plug_data(wf, mtcars)
wf

drop_data(wf)

replace_data(wf, iris)

}
