% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-action-formula.R
\name{plug_formula}
\alias{plug_formula}
\alias{drop_formula}
\alias{replace_formula}
\title{Add formula terms to a tidyflow}
\usage{
plug_formula(x, formula, ..., blueprint = NULL)

drop_formula(x)

replace_formula(x, formula, ..., blueprint = NULL)
}
\arguments{
\item{x}{A tidyflow}

\item{formula}{A formula specifying the terms of the model. It is advised to
not do preprocessing in the formula, and instead use a recipe if that is
required.}

\item{...}{Not used.}

\item{blueprint}{A hardhat blueprint used for fine tuning the preprocessing.
If \code{NULL}, \code{\link[hardhat:default_formula_blueprint]{hardhat::default_formula_blueprint()}} is used.}
}
\value{
The tidyflow \code{x}, updated with either a new or removed formula preprocessor.
}
\description{
\itemize{
\item \code{plug_formula()} specifies the terms of the model through the usage of a
formula.
\item \code{drop_formula()} removes the formula as well as any downstream objects
that might get created after the formula is used for preprocessing, such as
terms. Additionally, if the model has already been fit, then the fit is
removed.
\item \code{replace_formula()} first removes the formula, then replaces the previous
formula with the new one. Any model that has already been fit based on this
formula will need to be refit.
}
}
\details{
To fit a tidyflow, one of \code{plug_formula()} or \code{plug_recipe()} \emph{must} be
specified, but not both.

By default \code{tidyflow} leaves \code{workflows} to figure out which
type of factor transformation to happen (either leave factor as is,
transform to N-1 dummies or use a one-hot encoding approach of N dummy columns).
These transformations depend on the specific model supplied in \code{plug_model}.
See \code{\link[workflows]{add_formula}} for more details on how transformations
are handled.

However, \code{plug_formula} allows to override the type of transformation
using the blueprint. For example, by passing
\code{default_formula_blueprint(intercept = TRUE, indicators = "none")} to
the blueprint argument of \code{plug_formula} you can
enforce that the factor is left without transforming. You can also use
\code{default_formula_blueprint(intercept = TRUE, indicators = "traditional")} and
\code{default_formula_blueprint(intercept = TRUE, indicators = "one_hot")} to
transform factors to N-1 dummies or N dummies respectively.

For example, to transform all factors to one-hot encoding, you can pass
the blueprint to \code{plug_formula}:\if{html}{\out{<div class="r">}}\preformatted{bp <- default_formula_blueprint(intercept = TRUE, indicators = "one_hot")
iris \%>\%
  tidyflow(seed = 21315)
  plug_formula(Sepal.Length ~ Species, blueprint = bp) \%>\%
  plug_model(set_engine(parsnip::linear_reg(), "lm")) \%>\%
  fit()
}\if{html}{\out{</div>}}
}
\examples{

# Just for the pipe: \%>\%
library(tibble)

tflow <-
  mtcars \%>\%
  tidyflow(seed = 652341) \%>\% 
  plug_formula(mpg ~ .)

tflow

drop_formula(tflow)

replace_formula(tflow, mpg ~ disp)
}
