% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-action-split.R
\name{plug_split}
\alias{plug_split}
\alias{drop_split}
\alias{replace_split}
\title{Add a split specification to a tidyflow}
\usage{
plug_split(x, .f, ...)

drop_split(x)

replace_split(x, .f, ...)
}
\arguments{
\item{x}{A tidyflow}

\item{.f}{A function to be applied to the dataset in the tidyflow. Must
return an object of class \code{rsplit}. See package
\code{\link[rsample]{rsample}}.}

\item{...}{arguments passed to \code{.f}. These arguments must be named.
The processing of \code{...} respects the quotation rules from \code{.f}.
In other words, if the function allows variables as strings \strong{and}
as names, the user can specify both. See the example sections.}
}
\value{
\code{x}, updated with either a new or removed split specification.
}
\description{
\itemize{
\item \code{plug_split()} specifies the type of splitting used in the analysis. It
accepts a function \code{.f} that will be applied to the data. Only
functions which return an \code{rsplit} object will be allowed. See
package \code{\link[rsample]{rsample}} and the details section. If a model
has been fit before adding the split, it will need to be refit.
\item \code{drop_split()} removes the split specification from the tidyflow. Note
that it keeps other preprocessing steps such as the recipe.
\item \code{replace_split()} first removes the split, then adds a new split
specification. Any model that has already been fit based on this
split will need to be refit.
}
}
\details{
The split specification is an optional step in the tidyflow. You can add a
dataframe, prepare a recipe and fit the model without splitting into
training/testing.

When applied to the data, the function \code{.f} must return an object
of class \code{rsplit}. These are functions which come from the
\code{\link[rsample]{rsample}} package such as
\code{\link[rsample]{initial_split}}.
}
\examples{
library(tibble)
library(rsample)

wf <-
 mtcars \%>\%
 tidyflow() \%>\%
 plug_split(initial_split, prop = 0.8, strata = "cyl")

wf

# Strata as unquoted name
wf <- replace_split(wf, initial_split, prop = 0.8, strata = cyl)

wf

drop_split(wf)

# New split function
replace_split(wf, initial_time_split)

}
