% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{tidyflows-internals}
\alias{tidyflows-internals}
\alias{.fit_pre}
\alias{.fit_model}
\title{Internal tidyflow functions}
\usage{
.fit_pre(tflow)

.fit_model(tidyflow, control)
}
\arguments{
\item{tidyflow}{A tidyflow

For \code{.fit_pre()}, this should be a fresh tidyflow.

For \code{.fit_model()}, this should be a tidyflow that has already been trained
through \code{.fit_pre()}.}

\item{control}{A \code{\link[=control_tidyflow]{control_tidyflow()}} object}

\item{data}{A data frame of predictors and outcomes to use when fitting the
tidyflow}
}
\description{
\code{.fit_pre()} and \code{.fit_model()} are internal tidyflow functions for
\emph{partially} fitting a tidyflow object. They are only exported for usage by
the tuning package, \href{https://github.com/tidymodels/tune}{tune}, and the
general user should never need to worry about them.
}
\examples{
library(parsnip)
library(recipes)

model <- linear_reg()
model <- set_engine(model, "lm")

base_tidyflow <- tidyflow(mtcars)
base_tidyflow <- plug_model(base_tidyflow, model)

formula_tidyflow <- plug_formula(base_tidyflow, mpg ~ cyl + log(disp))

partially_fit_tidyflow <- .fit_pre(formula_tidyflow)
fit_tidyflow <- .fit_model(partially_fit_tidyflow, control_tidyflow())
}
\keyword{internal}
