% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-raw_data.R
\name{add_data}
\alias{add_data}
\alias{remove_data}
\alias{update_data}
\title{Add a data to a tidyflow}
\usage{
add_data(x, data, ...)

remove_data(x)

update_data(x, data, ...)
}
\arguments{
\item{x}{A tidyflow}

\item{data}{A data frame or tibble.}

\item{...}{Not used.}
}
\value{
\code{x}, updated with either a new or removed data frame.
}
\description{
\itemize{
\item \code{add_data()} specifies the raw data used in the analysis. This data is used
for the preprocessing and fitting the model.
\item \code{remove_data()} removes the data from the tidyflow.
\item \code{update_data()} first removes the data, then replaces the previous
data with the new one. Any model that has already been fit based on this
data will need to be refit.
}
}
\details{
To fit a tidyflow, \code{data} \emph{must} be specified.
}
\examples{

wf <- tidyflow()
wf <- add_data(wf, mtcars)
wf

remove_data(wf)

update_data(wf, iris)

}
