% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit-tidyflow}
\alias{fit-tidyflow}
\alias{fit.tidyflow}
\title{Fit a tidyflow object}
\usage{
\method{fit}{tidyflow}(tflow, ..., control = control_tidyflow())
}
\arguments{
\item{tflow}{A tidyflow object}

\item{...}{Not used}

\item{control}{A \code{\link[=control_tidyflow]{control_tidyflow()}} object}
}
\value{
If a tuning grid has been specified with \code{\link{plug_grid}}, then the
result is a data frame with the tuning results. Can be extracted with
\code{\link{pull_tflow_fit_tuning}}. If no tuning grid has been specified,
the result is a parsnip model. It can be extracted with
\code{\link{pull_tflow_fit}}.
}
\description{
Fitting a tidyflow currently involves two main steps:
\itemize{
\item Preprocessing the data using a formula preprocessor, or by calling
\code{\link[recipes:prep]{recipes::prep()}} on a recipe.
\item Fitting the underlying parsnip model using \code{\link[parsnip:fit.model_spec]{parsnip::fit.model_spec()}}.
}
}
\details{
In the future, there will also be \emph{postprocessing} steps that can be added
after the model has been fit.
}
\examples{
library(parsnip)
library(recipes)
library(tune)
library(dials)
library(rsample)

# Fit a simple linear model
model <- linear_reg()
model <- set_engine(model, "lm")

formula_tidyflow <-
 mtcars \%>\%
 tidyflow() \%>\%
 plug_formula(mpg ~ cyl + log(disp)) \%>\%
 plug_model(model)

fit(formula_tidyflow)

recipe_tidyflow <-
 formula_tidyflow \%>\%
 drop_formula() \%>\% 
 plug_recipe(~ recipe(mpg ~ cyl + disp, .x) \%>\% step_log(disp))

fit(recipe_tidyflow)

# Fit a regularized regression through a grid search.
# Do this by updating the already defined model:
new_mod <- set_engine(linear_reg(penalty = tune(), mixture = tune()),
                      "glmnet")
tuned_res <-
 formula_tidyflow \%>\%
  plug_resample(vfold_cv, v = 2) \%>\% 
  replace_formula(mpg ~ .) \%>\% 
  replace_model(new_mod) \%>\%
  plug_grid(grid_regular, levels = 2) \%>\%
  fit()

# Extract the tuning fit:
pull_tflow_fit_tuning(tuned_res)

# TODO: Extract best params and refit once complete_tflow is ready
 

}
