% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyflow-package.R
\docType{package}
\name{tidyflow-package}
\alias{tidyflow-package}
\alias{_PACKAGE}
\title{tidyflow: Modeling Tidyflows}
\description{
Managing a modelling flow such as data, training/testing splitting, resampling, preprocessing and statistical modelling, can often be challenging. The goal of 'tidyflow' is to streamline this process by bundling your tidyflow within the same object.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cimentadaj/tidyflow}
  \item Report bugs at \url{https://github.com/cimentadaj/tidyflow/issues}
}

}
\author{
\strong{Maintainer}: Jorge Cimentada \email{cimentadaj@gmail.com}

Authors:
\itemize{
  \item Davis Vaughan \email{davis@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
